Move assignment of actionCollection to fix startup crash
Starting KTimeTracker results in an immediate crash. This is due to a
use before initialization bug. The bug exists since 5e978ac9, where
actions have been surrounded by blocks, but also the assignment of
`m_actionCollection` has moved below the actions.

The member `m_actionCollection` has to be assigned before
`slotUpdateButtons()` is called, because it will be used there (in
`action`). This commit fixes this by moving it a few lines up.